refactor: Layout 컴포넌트 templates으로 이동
