+++ src/pages/HomePage/HomePage.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
-import Layout from '../../components/common/Layout';
 import LocalApiErrorBoundary from '../../components/errors/LocalApiErrorBoundary';
+import Layout from '../../components/templates/Layout';
 import OutfitFeedSkeleton from '../../features/outfits/components/OutfitFeed/OutfitFeedSkeleton';
 import OutfitFilter from '../../features/outfits/components/OutfitFilter';
 import HomeOutfitFeed from './components/HomeOutfitFeed';
